requirements.txt
asgiref==3.7.2
Django==3.2
django-crispy-forms==2.0
gunicorn==20.0.4
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
python-slugify
typing_extensions==4.6.3
whitenoise==6.4.0
